# Performances: snapshot socket: use a single bpf iterator