--- a/pkg/gadgets/snapshot/socket/tracer/tracer.go
+++ b/pkg/gadgets/snapshot/socket/tracer/tracer.go
@@ -19,32 +19,29 @@ package tracer
 import (
 	"bufio"
 	"encoding/binary"
+	"errors"
 	"fmt"
 	"net"
+	"strings"
 
+	"github.com/cilium/ebpf"
 	"github.com/cilium/ebpf/link"
 
-	containercollection "github.com/inspektor-gadget/inspektor-gadget/pkg/container-collection"
-	containerutils "github.com/inspektor-gadget/inspektor-gadget/pkg/container-utils"
 	"github.com/inspektor-gadget/inspektor-gadget/pkg/gadgets"
 	socketcollectortypes "github.com/inspektor-gadget/inspektor-gadget/pkg/gadgets/snapshot/socket/types"
-	"github.com/inspektor-gadget/inspektor-gadget/pkg/netnsenter"
+	"github.com/inspektor-gadget/inspektor-gadget/pkg/kallsyms"
+	"github.com/inspektor-gadget/inspektor-gadget/pkg/tcpbits"
 	eventtypes "github.com/inspektor-gadget/inspektor-gadget/pkg/types"
 )
 
-//go:generate go run github.com/cilium/ebpf/cmd/bpf2go -target bpfel -cc clang iterTCPv4 ./bpf/tcp4-collector.c -- -I../../../../${TARGET} -Werror -O2 -g -c -x c
-//go:generate go run github.com/cilium/ebpf/cmd/bpf2go -target bpfel -cc clang iterUDPv4 ./bpf/udp4-collector.c -- -I../../../../${TARGET} -Werror -O2 -g -c -x c
+//go:generate go run github.com/cilium/ebpf/cmd/bpf2go -target bpfel -cc clang iterSockets ./bpf/sockets.c -- -I../../../../${TARGET} -I ../../../common/ -Werror -O2 -g -c -x c
 
 type Tracer struct {
-	iters map[socketcollectortypes.Proto]*link.Iter
+	iter       *link.Iter
+	MountnsMap *ebpf.Map
 
-	// visitedNamespaces is a map where the key is the netns inode number and
-	// the value is the pid of one of the containers that share that netns. Such
-	// pid is used by NetnsEnter. TODO: Improve NetnsEnter to also work with the
-	// netns directly.
-	visitedNamespaces map[uint64]uint32
-	protocols         socketcollectortypes.Proto
-	eventHandler      func([]*socketcollectortypes.Event)
+	protocols    socketcollectortypes.Proto
+	eventHandler func([]*socketcollectortypes.Event)
 }
 
 func parseIPv4(ipU32 uint32) string {
@@ -60,17 +57,7 @@ func parseIPv4(ipU32 uint32) string {
 
 // Format from socket_bpf_seq_print() in bpf/socket_common.h
 func parseStatus(proto string, statusUint uint8) (string, error) {
-	statusMap := [...]string{
-		"ESTABLISHED", "SYN_SENT", "SYN_RECV",
-		"FIN_WAIT1", "FIN_WAIT2", "TIME_WAIT", "CLOSE", "CLOSE_WAIT",
-		"LAST_ACK", "LISTEN", "CLOSING", "NEW_SYN_RECV",
-	}
-
-	// Kernel enum starts from 1, adjust it to the statusMap
-	if statusUint == 0 || len(statusMap) <= int(statusUint-1) {
-		return "", fmt.Errorf("invalid %s status: %d", proto, statusUint)
-	}
-	status := statusMap[statusUint-1]
+	status := tcpbits.TCPState(statusUint)
 
 	// Transform TCP status into something more suitable for UDP
 	if proto == "UDP" {
@@ -87,112 +74,144 @@ func parseStatus(proto string, statusUint uint8) (string, error) {
 	return status, nil
 }
 
-func getTCPIter() (*link.Iter, error) {
-	objs := iterTCPv4Objects{}
-	if err := loadIterTCPv4Objects(&objs, nil); err != nil {
-		return nil, fmt.Errorf("failed to load TCP BPF objects: %w", err)
+func (t *Tracer) loadIter() error {
+	spec, err := loadIterSockets()
+	if err != nil {
+		return fmt.Errorf("load sockets BPF programs: %w", err)
+	}
+
+	mapReplacements := map[string]*ebpf.Map{}
+	filterByMntNs := false
+
+	if t.MountnsMap != nil {
+		filterByMntNs = true
+		mapReplacements[gadgets.MntNsFilterMapName] = t.MountnsMap
+	}
+
+	consts := map[string]interface{}{
+		"skip_tcp":                t.protocols != socketcollectortypes.ALL && t.protocols != socketcollectortypes.TCP,
+		"skip_udp":                t.protocols != socketcollectortypes.ALL && t.protocols != socketcollectortypes.UDP,
+		gadgets.FilterByMntNsName: filterByMntNs,
+	}
+
+	if err := spec.RewriteConstants(consts); err != nil {
+		return fmt.Errorf("RewriteConstants: %w", err)
+	}
+
+	kallsyms.SpecUpdateAddresses(spec, []string{"socket_file_ops"})
+
+	opts := ebpf.CollectionOptions{
+		MapReplacements: mapReplacements,
+	}
+	objs := &iterSocketsObjects{}
+	if err := spec.LoadAndAssign(objs, &opts); err != nil {
+		var errVerifier *ebpf.VerifierError
+		if errors.As(err, &errVerifier) {
+			fmt.Printf("Error: %+v\n", errVerifier)
+		}
+		return fmt.Errorf("load sockets BPF iterator: %w", err)
 	}
 	defer objs.Close()
 
-	it, err := link.AttachIter(link.IterOptions{
-		Program: objs.IgSnapTcp4,
+	t.iter, err = link.AttachIter(link.IterOptions{
+		Program: objs.IgSocketsIt,
 	})
 	if err != nil {
-		return nil, fmt.Errorf("failed to attach TCP BPF iterator: %w", err)
+		return fmt.Errorf("attach sockets BPF iterator: %w", err)
 	}
 
-	return it, nil
-}
-
-func getUDPIter() (*link.Iter, error) {
-	objs := iterUDPv4Objects{}
-	if err := loadIterUDPv4Objects(&objs, nil); err != nil {
-		return nil, fmt.Errorf("failed to load UDP BPF objects: %w", err)
-	}
-	defer objs.Close()
-
-	it, err := link.AttachIter(link.IterOptions{
-		Program: objs.IgSnapUdp4,
-	})
-	if err != nil {
-		return nil, fmt.Errorf("failed to attach UDP BPF iterator: %w", err)
-	}
-
-	return it, nil
+	return nil
 }
 
 // RunCollector is currently exported so it can be called from Collect()
-func (t *Tracer) RunCollector(pid uint32, podname, namespace, node string) ([]*socketcollectortypes.Event, error) {
+func (t *Tracer) RunCollector() ([]*socketcollectortypes.Event, error) {
 	sockets := []*socketcollectortypes.Event{}
-	err := netnsenter.NetnsEnter(int(pid), func() error {
-		for iterKey, it := range t.iters {
-			if t.protocols != socketcollectortypes.ALL && t.protocols != iterKey {
-				continue
-			}
 
-			reader, err := it.Open()
-			if err != nil {
-				return fmt.Errorf("failed to open BPF iterator: %w", err)
-			}
-			defer reader.Close()
+	type socketKey struct {
+		proto string
+		netns uint32
+		inode uint64
+	}
+	seenSocket := map[socketKey]bool{}
 
-			scanner := bufio.NewScanner(reader)
-			for scanner.Scan() {
-				var status, proto string
-				var destp, srcp uint16
-				var dest, src uint32
-				var hexStatus uint8
-				var inodeNumber uint64
-
-				// Format from socket_bpf_seq_print() in bpf/socket_common.h
-				// IP addresses and ports are in host-byte order
-				len, err := fmt.Sscanf(scanner.Text(), "%s %08X %04X %08X %04X %02X %d",
-					&proto, &src, &srcp, &dest, &destp, &hexStatus, &inodeNumber)
-				if err != nil || len != 7 {
-					return fmt.Errorf("failed to parse sockets information: %w", err)
-				}
-
-				status, err = parseStatus(proto, hexStatus)
-				if err != nil {
-					return err
-				}
-
-				// TODO: Receive the netns from caller
-				netns, err := containerutils.GetNetNs(int(pid))
-				if err != nil {
-					return fmt.Errorf("getting netns for pid %d: %w", pid, err)
-				}
-
-				sockets = append(sockets, &socketcollectortypes.Event{
-					Event: eventtypes.Event{
-						Type: eventtypes.NORMAL,
-						// TODO: This can be removed as events will be enriched
-						//  by the eventHandler
-						CommonData: eventtypes.CommonData{
-							Node:      node,
-							Namespace: namespace,
-							Pod:       podname,
-						},
-					},
-					Protocol:      proto,
-					LocalAddress:  parseIPv4(src),
-					LocalPort:     srcp,
-					RemoteAddress: parseIPv4(dest),
-					RemotePort:    destp,
-					Status:        status,
-					InodeNumber:   inodeNumber,
-					WithNetNsID:   eventtypes.WithNetNsID{NetNsID: netns},
-				})
-			}
-
-			if err := scanner.Err(); err != nil {
-				return fmt.Errorf("failed reading output of BPF iterator: %w", err)
-			}
-		}
-		return nil
-	})
+	reader, err := t.iter.Open()
 	if err != nil {
-		return nil, err
+		return nil, fmt.Errorf("open BPF iterator: %w", err)
+	}
+	defer reader.Close()
+
+	scanner := bufio.NewScanner(reader)
+	for scanner.Scan() {
+		var status, proto string
+		var ipversion int
+		var destp, srcp uint16
+		var dest, src string
+		var hexStatus uint8
+		var inodeNumber uint64
+		var mntns uint64
+		var netns uint32
+		var command string
+		var pid, parentPid int
+		var uid, gid uint32
+
+		// Format from socket_bpf_seq_print() in bpf/socket_common.h
+		// IP addresses and ports are in host-byte order
+		text := scanner.Text()
+		matchedElems, err := fmt.Sscanf(text, "%s %d %s %s %04X %04X %02X %d %d %d %d %d %d %d",
+			&proto, &ipversion,
+			&src, &dest, &srcp, &destp,
+			&hexStatus, &inodeNumber, &netns, &mntns,
+			&parentPid, &pid, &uid, &gid)
+		if err != nil {
+			return nil, fmt.Errorf("parse sockets information: %w", err)
+		}
+		if matchedElems != 14 {
+			return nil, fmt.Errorf("parse sockets information: found %d fields", matchedElems)
+		}
+		textSplit := strings.SplitN(text, " ", 15)
+		if len(textSplit) != 15 {
+			return nil, fmt.Errorf("failed to parse process information, expected 15 matched elements had %d", len(textSplit))
+		}
+		command = textSplit[14]
+
+		status, err = parseStatus(proto, hexStatus)
+		if err != nil {
+			return nil, err
+		}
+
+		// If two processes share the same file descriptor table (see CLONE_FILES in man 2 clone),
+		// the same socket will be seen twice. This is the case with nginx worker processes.
+		// Skip sockets that have already been seen.
+		key := socketKey{proto, netns, inodeNumber}
+		if seenSocket[key] {
+			continue
+		}
+		seenSocket[key] = true
+
+		sockets = append(sockets, &socketcollectortypes.Event{
+			Event: eventtypes.Event{
+				Type: eventtypes.NORMAL,
+			},
+			WithMountNsID: eventtypes.WithMountNsID{MountNsID: mntns},
+			WithNetNsID:   eventtypes.WithNetNsID{NetNsID: uint64(netns)},
+			Pid:           pid,
+			Uid:           uid,
+			Gid:           gid,
+			Command:       command,
+			ParentPid:     parentPid,
+			Protocol:      proto,
+			IPVersion:     ipversion,
+			LocalAddress:  net.ParseIP(src).String(),
+			LocalPort:     srcp,
+			RemoteAddress: net.ParseIP(dest).String(),
+			RemotePort:    destp,
+			Status:        status,
+			InodeNumber:   inodeNumber,
+		})
+	}
+
+	if err := scanner.Err(); err != nil {
+		return nil, fmt.Errorf("reading output of BPF iterator: %w", err)
 	}
 
 	return sockets, nil
@@ -202,13 +221,11 @@ func (t *Tracer) RunCollector(pid uint32, podname, namespace, node string) ([]*s
 
 func NewTracer(protocols socketcollectortypes.Proto) (*Tracer, error) {
 	tracer := &Tracer{
-		visitedNamespaces: make(map[uint64]uint32),
-		protocols:         protocols,
-		iters:             make(map[socketcollectortypes.Proto]*link.Iter),
+		protocols: protocols,
 	}
 
-	if err := tracer.openIters(); err != nil {
-		tracer.CloseIters()
+	if err := tracer.loadIter(); err != nil {
+		tracer.CloseIter()
 		return nil, fmt.Errorf("installing tracer: %w", err)
 	}
 
@@ -216,22 +233,7 @@ func NewTracer(protocols socketcollectortypes.Proto) (*Tracer, error) {
 }
 
 func (g *GadgetDesc) NewInstance() (gadgets.Gadget, error) {
-	return &Tracer{
-		visitedNamespaces: make(map[uint64]uint32),
-		iters:             make(map[socketcollectortypes.Proto]*link.Iter),
-	}, nil
-}
-
-func (t *Tracer) AttachContainer(container *containercollection.Container) error {
-	if _, ok := t.visitedNamespaces[container.Netns]; ok {
-		return nil
-	}
-	t.visitedNamespaces[container.Netns] = container.Pid
-	return nil
-}
-
-func (t *Tracer) DetachContainer(container *containercollection.Container) error {
-	return nil
+	return &Tracer{}, nil
 }
 
 func (t *Tracer) SetEventHandlerArray(handler any) {
@@ -242,58 +244,37 @@ func (t *Tracer) SetEventHandlerArray(handler any) {
 	t.eventHandler = nh
 }
 
-// CloseIters is currently exported so it can be called from Collect()
-func (t *Tracer) CloseIters() {
-	for _, it := range t.iters {
-		it.Close()
+// CloseIter is currently exported so it can be called from Collect()
+func (t *Tracer) CloseIter() {
+	if t.iter != nil {
+		t.iter.Close()
 	}
-	t.iters = nil
-}
-
-func (t *Tracer) openIters() error {
-	var it *link.Iter
-	var err error
-
-	if t.protocols == socketcollectortypes.TCP || t.protocols == socketcollectortypes.ALL {
-		it, err = getTCPIter()
-		if err != nil {
-			return err
-		}
-		t.iters[socketcollectortypes.TCP] = it
-	}
-
-	if t.protocols == socketcollectortypes.UDP || t.protocols == socketcollectortypes.ALL {
-		it, err = getUDPIter()
-		if err != nil {
-			return err
-		}
-		t.iters[socketcollectortypes.UDP] = it
-	}
-
-	return nil
+	t.iter = nil
 }
 
 func (t *Tracer) Run(gadgetCtx gadgets.GadgetContext) error {
 	protocols := gadgetCtx.GadgetParams().Get(ParamProto).AsString()
 	t.protocols, _ = socketcollectortypes.ProtocolsMap[protocols]
 
-	defer t.CloseIters()
-	if err := t.openIters(); err != nil {
+	defer t.CloseIter()
+	if err := t.loadIter(); err != nil {
 		return fmt.Errorf("installing tracer: %w", err)
 	}
 
 	allSockets := []*socketcollectortypes.Event{}
-	for netns, pid := range t.visitedNamespaces {
-		// TODO: Remove podname, namespace and node arguments from RunCollector.
-		// The enrichment will be done in the event handler. In addition, pass
-		// the netns to avoid retrieving it again in RunCollector.
-		sockets, err := t.RunCollector(pid, "", "", "")
-		if err != nil {
-			return fmt.Errorf("snapshotting sockets in netns %d: %w", netns, err)
-		}
-		allSockets = append(allSockets, sockets...)
+	// TODO: Remove podname, namespace and node arguments from RunCollector.
+	// The enrichment will be done in the event handler. In addition, pass
+	// the netns to avoid retrieving it again in RunCollector.
+	sockets, err := t.RunCollector()
+	if err != nil {
+		return fmt.Errorf("read sockets: %w", err)
 	}
+	allSockets = append(allSockets, sockets...)
 
 	t.eventHandler(allSockets)
 	return nil
+}
+
+func (t *Tracer) SetMountNsMap(mountnsMap *ebpf.Map) {
+	t.MountnsMap = mountnsMap
 }