… docs/gadgets/snapshot/socket.md
… pkg/gadgets/snapshot/socket/tracer/bpf/socket-common.h
… pkg/gadgets/snapshot/socket/tracer/bpf/tcp4-collector.c
… pkg/gadgets/snapshot/socket/tracer/bpf/udp4-collector.c
… pkg/gadgets/snapshot/socket/tracer/itersockets_bpfel.go
… pkg/gadgets/snapshot/socket/tracer/itersockets_bpfel.o
… pkg/gadgets/snapshot/socket/tracer/itertcpv4_bpfel.go
… pkg/gadgets/snapshot/socket/tracer/itertcpv4_bpfel.o
… pkg/gadgets/snapshot/socket/tracer/iterudpv4_bpfel.go
… pkg/gadgets/snapshot/socket/tracer/iterudpv4_bpfel.o