# snapshot socket: use a single bpf iterator

The snapshot socket gadget was previously using bpf iterators of types iter/udp and iter/tcp. Those bpf iterators need to be executed in each network namespace. This causes bad performance when there are a lot of containers on the node.

This patch uses a bpf iterator of type iter/task_file, iterating on all file descriptors from all processes and skipping the ones that are not tcp or udp sockets.

Benchmarks show performance improvement by 50%:

```
goos: linux
goarch: amd64
pkg: github.com/inspektor-gadget/inspektor-gadget/internal/benchmarks
cpu: Intel(R) Core(TM) i7-6500U CPU @ 2.50GHz
                                                        │  main.bench  │            patched.bench             │
                                                        │    sec/op    │    sec/op     vs base                │
AllGadgetsWithContainers/container100/snapshot-socket-4   516.9m ± 26%   254.4m ± 27%  -50.79% (p=0.000 n=10)
```

The benchmark was produced using the method described in the documentation:
https://github.com/inspektor-gadget/inspektor-gadget/blob/main/docs/devel/CONTRIBUTING.md#benchmarks

Using iter/task_file has the benefit to bring process enrichment (pid, comm, etc).

But it only iterates on sockets that are installed in a process:
1. tcp sockets in TIME_WAIT state are not visible anymore because those are not associated with a struct sock and they do not have an underlying file. See [tcp.h#L442](https://elixir.bootlin.com/linux/v5.15.12/source/include/linux/tcp.h#L442) and [inet_timewait_sock.h#L33](https://elixir.bootlin.com/linux/v5.15.12/source/include/net/inet_timewait_sock.h#L33)
2. server-side tcp sockets that are not accepted yet (the server didn't call the `accept()` system call yet) are no longer visible since they are not a file descriptor (yet).

- [x] Rebased after #1593 got merged in order to reuse the kallsyms package.

## How to use

No changes.

## Testing done

- [x] ig: with tcp sockets using netcat in a Docker container
- [x] ig: udp sockets
- [x] kubectl-gadget
- [x] IPv6

## Issues

Fixes https://github.com/inspektor-gadget/inspektor-gadget/issues/1611
Fixes https://github.com/inspektor-gadget/inspektor-gadget/issues/744