@mauriciovasquezbernal 2023-05-09
> tcp sockets in TIME_WAIT state are not visible anymore because those are not associated with a struct sock and they do not have an underlying file. See [tcp.h#L442](https://elixir.bootlin.com/linux/v5.15.12/source/include/linux/tcp.h#L442) and [inet_timewait_sock.h#L33](https://elixir.bootlin.com/linux/v5.15.12/source/include/net/inet_timewait_sock.h#L33)

Is it really fine not to show sockets on `TIME_WAIT` state? It seems to me that information is useful and not having them will be a shame.