@alban 2023-05-11
> Is it really fine not to show sockets on TIME_WAIT state? It seems to me that information is useful and not having them will be a shame.

In that case, using an "iter/task_file" does not work, and we cannot enrich the events with container and process data. I suggest the following solution:
1. Close this PR.
2. Try to implement the snapshot socket with reusing the same instances of ebpf programs. It will help a bit with performance (bpf verifier executed only one time per ebpf program) but we will still have to enter each different network namespace in Go and open/read the iterator each time. It will not have the container and process enrichment.
3. Create a new gadget "snapshot file" using an "iter/task_file", iterating on all open files, including sockets, but not the TIME_WAIT & incoming-but-not-yet-accepted TCP sockets since those don't have a file descriptor. This gadget can also show regular files, and other kinds of files (bpf programs, pipes, etc). It will have container and process enrichment.

In this way, users can have both process enrichment and non-file sockets with good performances but in two different gadgets.

What do you think?